Refactor cli command layout
Having each subcommand defined similarly will hopefully make the
common root level flag overrides easier to implement.
