+++ .gitignore
@@ -9,3 +9,4 @@ release_notes.md
 .idea
 .vscode
 .op
+benthos
